Resolved critical performance regression in the lightmapper.
Still a major deadlock/stutter bug occurring after some threading changes. Will tackle this next.

Refactored VoxelData and VoxelColor to be more ergonomic. Color was specifically modified as part of fixing the lightmapper and now uses a LUT for preserving color fidelity with 16-bit color.

Had Claude come up with a more thorough plan to refactor the chunk pipeline as it has become a bit of a mess. Will begin execution on this as it will likely clean up the existing threading bug as well.